Reduce integer precision to lower mem use

+++ src/branch_sequences.c
@@ -1022,17 +1022,17 @@ double get_block_likelihood(int branch_genome_size, int number_of_branch_snps, i
 
 int calculate_genome_length_excluding_blocks_and_gaps(char * sequence, int length_of_sequence, int ** block_coordinates, int num_blocks)
 {
-	int * bases_to_be_excluded;  
-	bases_to_be_excluded = (int*) calloc((length_of_sequence + 1),sizeof(int));
+    int8_t* bases_to_be_excluded;
+	bases_to_be_excluded = (int8_t*) calloc((length_of_sequence + 1),sizeof(int8_t));
 	
-  int genome_length = length_of_sequence;
+    int genome_length = length_of_sequence;
 	int i = 0;
 	for(i = 0; i<length_of_sequence; i++)
 	{
 		if(sequence[i] == 'N' || sequence[i] == '-' )
 		{
 			bases_to_be_excluded[i] = 1;
-      genome_length--;
+            genome_length--;
 		}
 	}
 	
@@ -1048,11 +1048,11 @@ int calculate_genome_length_excluding_blocks_and_gaps(char * sequence, int lengt
 		int block_index = 0;
 		for(block_index = block_coordinates[0][j]; block_index <= block_coordinates[1][j]; block_index++ )
 		{
-      if(bases_to_be_excluded[block_index-1] == 0)
-      {
-        bases_to_be_excluded[block_index-1] = 1;
-        genome_length--;
-      }
+          if(bases_to_be_excluded[block_index-1] == 0)
+          {
+            bases_to_be_excluded[block_index-1] = 1;
+            genome_length--;
+          }
 		}
 	}
